[API] Final refactor (route/controller & endpoints)
- All API logic has been split from the endpoint definitions.
- Endpoints are defined in `.routes` files.
- API logic is defined in `.controller` files.
- Endpoints have been put into more logical routes, e.g; all wallet APIs are under `/wallet/`, all activity APIs are under `/activity/`, etc.
- Backwards-compat with prior wallet APIs has been kept for v1.1.5, scheduled to be removed in v1.1.6, giving one update of 'buffer' for services to upgrade smoothly.
